requirements.txt
fastapi
uvicorn
numpy
pandas
Pillow
matplotlib
scipy
